refactor(orchestrator): extract task delivery workflow
